[TASK] singled out virtual pass through command rewriter
